Jetty embedded. Ready for Match-class changes.
