+++ app/build.gradle
@@ -23,6 +23,11 @@ dependencies {
     // This dependency is used by the application.
     implementation 'com.google.guava:guava:29.0-jre'
     implementation 'org.postgresql:postgresql:42.2.18'
+    implementation 'org.postgresql:postgresql:42.2.18'
+    compile group: 'org.eclipse.jetty', name: 'jetty-server', version: '9.4.36.v20210114'
+    compile group: 'org.eclipse.jetty', name: 'jetty-servlet', version: '9.4.36.v20210114'
+    compile group: 'org.eclipse.jetty', name: 'jetty-webapp', version: '9.4.36.v20210114'
+    compile group: 'org.mortbay.jetty', name: 'servlet-api', version: '3.0.20100224'
 }
 
 application {
