+++ app/src/main/java/kryaskov/app/game/ManualTeamBuilder.java
@@ -1,0 +1,27 @@
+package kryaskov.app.game;
+import java.util.ArrayList;
+
+public class ManualTeamBuilder extends TeamBuilder {
+    private Team team;
+
+    private PlayerBuilder builder;
+
+    private static int index = 1;
+
+    @Override
+    public Team createTeam() {
+        team = new Team();
+        team.setTeamName(UserInput.userInput.userTeamName(index));
+        createTeamSquad();
+        return team;
+    }
+
+    public void createTeamSquad() {
+        builder = UserInput.userInput.creationMethod(" игроков").equals("y") ? new AutomaticPlayerBuilder() : new ManualPlayerBuilder();
+        team.teamSquad = new ArrayList<>();
+
+        for (int i = 0; i < 6; i++) {
+            team.teamSquad.add(builder.createAPlayer());
+        }
+    }
+}
