+++ app/src/main/java/kryaskov/app/game/NewVolleyGame.java
@@ -1,0 +1,9 @@
+package kryaskov.app.game;
+
+public class NewVolleyGame {
+
+    public static void main(String[] args) {
+        Match.getMatch().go();
+    }
+}
+
