+++ app/src/main/java/kryaskov/app/game/PlayerBuilder.java
@@ -1,0 +1,5 @@
+package kryaskov.app.game;
+
+public abstract class PlayerBuilder {
+    public abstract Player createAPlayer();
+}
