+++ app/src/main/java/kryaskov/app/game/TeamBuilder.java
@@ -1,0 +1,5 @@
+package kryaskov.app.game;
+
+public abstract class TeamBuilder {
+    public abstract Team createTeam();
+}
