+++ app/src/main/java/kryaskov/app/servlets/GreetingServlet.java
@@ -1,0 +1,28 @@
+package kryaskov.app.servlets;
+
+import kryaskov.app.game.Match;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.PrintWriter;
+
+
+@WebServlet("/greeting")
+public class GreetingServlet extends HttpServlet {
+
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+        response.setContentType("text/html");
+        response.setCharacterEncoding("UTF-8");
+        response.setStatus(HttpServletResponse.SC_OK);
+        PrintWriter printWriter = response.getWriter();
+        printWriter.write("<p>Hello!</p>" + "<a href='options'>Press here to launch the best Volleyball Simulator</a>");
+        printWriter.close();
+
+    }
+}
