+++ app/src/main/java/kryaskov/app/servlets/ManualTeamServlet.java
@@ -1,0 +1,25 @@
+package kryaskov.app.servlets;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.PrintWriter;
+
+@WebServlet("manual")
+public class ManualTeamServlet extends HttpServlet {
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("text/html");
+        response.setStatus(HttpServletResponse.SC_OK);
+
+        PrintWriter printWriter = response.getWriter();
+        printWriter.write("Here will be possible to create team manually");
+
+        printWriter.close();
+
+    }
+
+}
