+++ app/src/main/java/kryaskov/app/servlets/PlayServlet.java
@@ -1,0 +1,27 @@
+package kryaskov.app.servlets;
+
+import kryaskov.app.game.Match;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.PrintWriter;
+
+@WebServlet("/play")
+public class PlayServlet extends HttpServlet {
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("text/html");
+        response.setStatus(HttpServletResponse.SC_OK);
+
+        PrintWriter printWriter = response.getWriter();
+        printWriter.write("The app will soon be ready!");
+
+
+        printWriter.close();
+        //Match.getMatch().go();
+    }
+}
