--- app/src/main/java/volley/NewVolleyGame.java
@@ -1,9 +1,0 @@
-package volley;
-
-public class NewVolleyGame {
-
-    public static void main(String[] args) {
-        Match.getMatch().go();
-    }
-}
-
