--- app/src/main/java/volley/PlayerBuilder.java
@@ -1,5 +1,0 @@
-package volley;
-
-public abstract class PlayerBuilder {
-    public abstract Player createAPlayer();
-}
