--- app/src/main/java/volley/Team.java
@@ -1,24 +1,0 @@
-package volley;
-import java.util.ArrayList;
-import java.util.List;
-
-class Team {
-    String teamName;
-    List<Player> teamSquad;
-
-    public void setTeamName(String gotTeamName) { teamName = gotTeamName; }
-
-    public String getTeamName() {
-        return teamName;
-    }
-
-    public List<Player> getTeamSquad() {
-        return teamSquad;
-    }
-
-    public void positioningChange() {
-        Player player = teamSquad.get(0);
-        teamSquad.remove(0);
-        teamSquad.add(player);
-    }
-}
