--- app/src/main/java/volley/TeamBuilder.java
@@ -1,5 +1,0 @@
-package volley;
-
-public abstract class TeamBuilder {
-    public abstract Team createTeam();
-}
