+++ app/src/main/webapp/WEB-INF/views/greeting.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Greeting</title>
+</head>
+<body>
+
+<h2 style="color: saddlebrown">Welcome to volleyball simulator!</h2>
+
+</body>
+</html>
